Split individual classes out of native-watcher-registry.js
