+++ lib/registry/registry-tree.js
@@ -1,0 +1,88 @@
+const {absolute} = require('./helper')
+
+const {RegistryNode} = require('./registry-node')
+const {RegistryWatcherNode} = require('./registry-watcher-node')
+
+// Private: Map userland filesystem watcher subscriptions efficiently to deliver filesystem change notifications to
+// each watcher with the most efficient coverage of native watchers.
+//
+// * If two watchers subscribe to the same directory, use a single native watcher for each.
+// * Re-use a native watcher watching a parent directory for a watcher on a child directory. If the parent directory
+//   watcher is removed, it will be split into child watchers.
+// * If any child directories already being watched, stop and replace them with a watcher on the parent directory.
+//
+// Uses a trie whose structure mirrors the directory structure.
+class RegistryTree {
+  // Private: Construct a tree with no native watchers.
+  //
+  // * `basePathSegments` the position of this tree's root relative to the filesystem's root as an {Array} of directory
+  //   names.
+  // * `createNative` {Function} used to construct new native watchers. It should accept an absolute path as an argument
+  //   and return a new {NativeWatcher}.
+  constructor (basePathSegments, createNative) {
+    this.basePathSegments = basePathSegments
+    this.root = new RegistryNode()
+    this.createNative = createNative
+  }
+
+  // Private: Identify the native watcher that should be used to produce events at a watched path, creating a new one
+  // if necessary.
+  //
+  // * `pathSegments` the path to watch represented as an {Array} of directory names relative to this {RegistryTree}'s
+  //   root.
+  // * `attachToNative` {Function} invoked with the appropriate native watcher and the absolute path to its watch root.
+  add (pathSegments, attachToNative) {
+    const absolutePathSegments = this.basePathSegments.concat(pathSegments)
+    const absolutePath = absolute(...absolutePathSegments)
+
+    const attachToNew = (childPaths) => {
+      const native = this.createNative(absolutePath)
+      const leaf = new RegistryWatcherNode(native, absolutePathSegments, childPaths)
+      this.root = this.root.insert(pathSegments, leaf)
+
+      const sub = native.onWillStop(split => {
+        sub.dispose()
+        const createNative = split ? this.createNative : null
+        this.root = this.root.remove(pathSegments, createNative) || new RegistryNode()
+      })
+
+      attachToNative(native, absolutePath)
+      return native
+    }
+
+    this.root.lookup(pathSegments).when({
+      parent: (parent, remaining) => {
+        // An existing NativeWatcher is watching the same directory or a parent directory of the requested path.
+        // Attach this Watcher to it as a filtering watcher and record it as a dependent child path.
+        const native = parent.getNativeWatcher()
+        parent.addChildPath(remaining)
+        attachToNative(native, absolute(...parent.getAbsolutePathSegments()))
+      },
+      children: children => {
+        // One or more NativeWatchers exist on child directories of the requested path. Create a new native watcher
+        // on the parent directory, note the subscribed child paths, and cleanly stop the child native watchers.
+        const newNative = attachToNew(children.map(child => child.path))
+
+        for (let i = 0; i < children.length; i++) {
+          const childNode = children[i].node
+          const childNative = childNode.getNativeWatcher()
+          childNative.reattachTo(newNative, absolutePath)
+          childNative.stop()
+        }
+      },
+      missing: () => attachToNew([])
+    })
+  }
+
+  // Private: Access the root node of the tree.
+  getRoot () {
+    return this.root
+  }
+
+  // Private: Return a {String} representation of this tree's structure for diagnostics and testing.
+  print () {
+    return this.root.print()
+  }
+}
+
+module.exports = {RegistryTree}
